compatibility with wx2.6 and vc7.1
git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@1648 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/gui/print_window.hpp
@@ -55,7 +55,7 @@ class PrintJob : public IntrusivePtrBase<PrintJob> {
 	
 	inline int num_pages() const {
 		int cards_per_page = max(1,layout.cards_per_page());
-		return (cards.size() + cards_per_page - 1) / cards_per_page;
+		return ((int)cards.size() + cards_per_page - 1) / cards_per_page;
 	}
 };
